Source/services/ingestion-service/requirements.txt
chromadb
pdfminer.six
pdf2image
pytesseract
Pillow
pandas
openpyxl
langdetect
